Moved sequence expansion logic to own module. No checks carried out for current functionality, will need addressing in the future to test and refine
